+++ 2D_game/src/data/main/FastEnemy.java
@@ -8,8 +8,8 @@ public class FastEnemy extends GameObject{
 	
 	private Handler handler;
 
-	public FastEnemy(int x, int y, ID id, Handler handler) {
-		super(x, y, id);
+	public FastEnemy(float x, float y, ID id, Handler handler) {
+		super((int)x, (int)y, id);
 		
 		this.handler = handler;
 		
@@ -18,7 +18,7 @@ public class FastEnemy extends GameObject{
 	}
 	
 	  public Rectangle getBounds() {
-		  return new Rectangle(x, y, 16, 16);
+		  return new Rectangle((int)x,(int) y, 16, 16);
 	  }
 
 
@@ -39,7 +39,7 @@ public class FastEnemy extends GameObject{
 	public void render(Graphics g) {
 		
 		g.setColor(Color.cyan);
-		g.fillRect(x, y, 16, 16);
+		g.fillRect((int)x, (int)y, 16, 16);
 		
 	}
