+++ 2D_game/src/data/main/Trail.java
@@ -17,8 +17,8 @@ public class Trail extends GameObject{
 	
 	//life = 0.01 - 0.1
 
-	public Trail(int x, int y, ID id, Color color, int width, int height, float life, Handler handler) {
-		super(x, y, id);
+	public Trail(float x, float y, ID id, Color color, int width, int height, float life, Handler handler) {
+		super((int)x, (int)y, id);
 		this.handler = handler;
 		this.color = color;
 		this.width = width;
@@ -38,7 +38,7 @@ public class Trail extends GameObject{
 		g2d.setComposite(makeTransparent(alpha));
 		
 		g.setColor(color);
-		g.fillRect(x, y, width, height);
+		g.fillRect((int)x,(int) y, width, height);
 		
 		g2d.setComposite(makeTransparent(1));
